[Update] : update entity - routine model entity의 최대 최소 총합별 무게 횟수 타입을 추가합니다. - workhistory entity를 추가합니다.

+++ core/model/src/main/java/com/gradation/lift/model/data/WorkSet.kt
@@ -6,6 +6,6 @@ weight : 무게
 repetition : 횟수
  */
 data class WorkSet(
-    val weight : Float,
+    val weight : Double,
     val repetition : Int
 )
